requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
bcrypt==4.0.1
PyJWT==2.8.0
PyMySQL==1.1.0
cryptography==41.0.1 